Patched wraparound detection.
Strangely, the wraparound calculation that we had did not work
reliably. It was counting too many wraps especially for z,
which of course is the axis where it hurts us most (the system can
mechanically lock up).

Now a more conservative wraparound counter. For fast movements we
might miss a turn, this needs to be manually checked.